requirements.txt
streamlit==1.30.0
openai==1.8.0
pinecone-client==3.0.1
python-dotenv==1.0.0
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.26.3
pandas==2.2.0
pdfplumber==0.10.3
langchain==0.1.0 